Add support for retrying 404 errors in GetData, which have started to appear from BCDA.
A bcda.ErrorRetryableHTTPStatus error is introduced to wrap retryable HTTP errors. ErrorUnauthorized is kept as is because it indicates that beyond just retrying, re-authentication is required (plus Go doesn't have good out of the box support for wrapping multiple errors).

In the future, we can consider introducing a custom error struct that holds onto structured information a caller may wish to interact with (e.g. holds the structured error codes, error body, etc). But at the moment may not be needed.

PiperOrigin-RevId: 461977015