Update ndjsonSink (and gcsNDJSONSink) to support non-blocking Write() calls with concurrent write processing. This helps performance by not blocking the pipeline on this I/O, since the pipeline processes each resource sequentially.
This also adds support for internal concurrent processing of writes within ndjsonSink across file shards. This may also mean that while fhir resources will be written to the same file sent by the server, they may be in a different order.

In practice, this second part--the internal concurrent processing portion over file shards--will not do much in the current usage pattern, because input ndjson files are processed sequentially in the pipeline and writes to each file shard must be sequential. However, if we support our own sharding mechanism in the future (which is reasonable to consider), this will yield gains.

PiperOrigin-RevId: 525168573