Move to our own sharding strategy for NDJSON sinks, and add more robust error handling in NDJSON sinks.
The new sharding simplifies the logic and allows for better throughput at the tradeoff of NOT storing resource types together in the same files. This is one of the optimizations I alluded to a prior change.

This also introduces better error handling where worker errors will not only be logged, they will also be retried (if it _might_ be loosely retriable), be recorded to a metric, and fail the Pipeline in Finalize().

Importantly, the sharing strategy will also allow us to Close files as they are filled instead of waiting until Finalize() to Close the files, which can lead to better performance in some file implementations as well, and less memory usage.

This change updates various tests, adds helpful test helpers, and ensures the GCS server is thread safe.

PiperOrigin-RevId: 539984464